# services/events-service-mock/python-flask-server/swagger_server/models/ticket_id.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class TicketID(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, tid: int=None):  # noqa: E501
        """TicketID - a model defined in Swagger

        :param tid: The tid of this TicketID.  # noqa: E501
        :type tid: int
        """
        self.swagger_types = {
            'tid': int
        }

        self.attribute_map = {
            'tid': 'tid'
        }

        self._tid = tid

    @classmethod
    def from_dict(cls, dikt) -> 'TicketID':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TicketID of this TicketID.  # noqa: E501
        :rtype: TicketID
        """
        return util.deserialize_model(dikt, cls)

    @property
    def tid(self) -> int:
        """Gets the tid of this TicketID.


        :return: The tid of this TicketID.
        :rtype: int
        """
        return self._tid

    @tid.setter
    def tid(self, tid: int):
        """Sets the tid of this TicketID.


        :param tid: The tid of this TicketID.
        :type tid: int
        """

        self._tid = tid