# services/events-service/python-flask-server/swagger_server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "Events API for ZeroWing Event Management System."
  version: "1.0.0"
  title: "Events API"
host: "localhost:8081"
basePath: "/v1"
schemes:
- "https"
- "http"
paths:
  /events:
    get:
      tags:
      - "Events"
      summary: "Retrieve a List of Events. Search by Event Title, Event category or\
        \ Event Description."
      operationId: "list_events"
      produces:
      - "application/json"
      parameters:
      - name: "event_title"
        in: "query"
        description: "The Event Title to search for."
        required: false
        type: "string"
      - name: "event_category"
        in: "query"
        description: "The Event Category to search for."
        required: false
        type: "string"
      - name: "event_desc"
        in: "query"
        description: "The Event Description to search for."
        required: false
        type: "string"
      - name: "host_id"
        in: "query"
        description: "The Host ID to search for."
        required: false
        type: "string"
      - name: "event_status"
        in: "query"
        description: "The Event Status to search for."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Events. Returns empty array if no Events\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/EventList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    post:
      tags:
      - "Events"
      summary: "Used to create an Event."
      operationId: "create_event"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Event object containing the Event details."
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "201":
          description: "Event created."
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    options:
      tags:
      - "Events"
      summary: "Used to respond to browser with Access-Control-Allow-Methods header.\
        \ Required for POST."
      operationId: "create_event_options"
      parameters: []
      responses:
        "200":
          description: "Response headers will indicate the allowable methods."
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
  /events/{event_id}:
    get:
      tags:
      - "Events"
      summary: "Retrieve Event details by Event ID."
      operationId: "get_event_details"
      produces:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be retrieved."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Event details successfully retrieved."
          schema:
            $ref: "#/definitions/Event"
        "404":
          description: "Event not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    put:
      tags:
      - "Events"
      summary: "Used to update the Event details. Replaces the Event resource."
      operationId: "update_event"
      consumes:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Event object to update. Performs a complete replace of the Event\
          \ details."
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "200":
          description: "Event successfully updated."
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "404":
          description: "Event not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    options:
      tags:
      - "Events"
      summary: "Used to respond to browser with Access-Control-Allow-Methods header.\
        \ Required for PUT."
      operationId: "update_event_options"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be updated."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Response headers will indicate the allowable methods."
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    patch:
      tags:
      - "Events"
      summary: "Used to update the Status of a single Event e.g. Cancel an Event."
      operationId: "update_event_status"
      consumes:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "The patch operation to perform. Only Event status update is\
          \ supported."
        required: true
        schema:
          $ref: "#/definitions/EventStatusUpdate"
      responses:
        "200":
          description: "Event status successfully updated."
          schema:
            $ref: "#/definitions/Event"
        "404":
          description: "Timeslot not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
  /venues:
    get:
      tags:
      - "Venues"
      summary: "Retrieve a List of Venues. Search by Venue Name."
      operationId: "list_venues"
      produces:
      - "application/json"
      parameters:
      - name: "venue_name"
        in: "query"
        description: "The Venue Name to search for."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Venues. Returns empty array if no Venues\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/VenueList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.venues_controller"
    post:
      tags:
      - "Venues"
      summary: "Used to create a Venue."
      operationId: "create_venue"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Venue object containing Venue details."
        required: true
        schema:
          $ref: "#/definitions/Venue"
      responses:
        "201":
          description: "Venue created."
          schema:
            $ref: "#/definitions/Venue"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.venues_controller"
    options:
      tags:
      - "Venues"
      summary: "Used to respond to browser with Access-Control-Allow-Methods header.\
        \ Required for POST."
      operationId: "create_venue_options"
      parameters: []
      responses:
        "200":
          description: "Response headers will indicate the allowable methods."
      x-swagger-router-controller: "swagger_server.controllers.venues_controller"
  /venues/{venue_id}:
    get:
      tags:
      - "Venues"
      summary: "Retrieve Venue details by Venue ID."
      operationId: "get_venue_details"
      produces:
      - "application/json"
      parameters:
      - name: "venue_id"
        in: "path"
        description: "ID of the Venue to be retrieved."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Venue details successfully retrieved."
          schema:
            $ref: "#/definitions/Venue"
        "404":
          description: "Event not found."
          schema:
            $ref: "#/definitions/VenueNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.venues_controller"
  /tickets:
    get:
      tags:
      - "Tickets"
      summary: "Retrieve a List of Tickets. Search by Event ID and Ticket Status."
      operationId: "list_tickets"
      produces:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "query"
        description: "The Event ID to search for."
        required: false
        type: "string"
      - name: "ticket_status"
        in: "query"
        description: "The Ticket Status to search for."
        required: false
        type: "string"
      - name: "booking_id"
        in: "query"
        description: "The Booking ID to search for."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Tickets. Returns empty array if no Tickets\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/TicketList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.tickets_controller"
  /bookings:
    get:
      tags:
      - "Bookings"
      summary: "Retrieve a List of Bookings. Search by Account ID and Booking Status."
      operationId: "list_bookings"
      produces:
      - "application/json"
      parameters:
      - name: "account_id"
        in: "query"
        description: "The Account ID to search for."
        required: false
        type: "string"
      - name: "booking_status"
        in: "query"
        description: "The Booking Status to search for."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Bookings. Returns empty array if no Bookings\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/BookingList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
    post:
      tags:
      - "Bookings"
      summary: "Used to create a Booking."
      operationId: "create_booking"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Booking object containing the Booking details."
        required: true
        schema:
          $ref: "#/definitions/Booking"
      responses:
        "201":
          description: "Booking created."
          schema:
            $ref: "#/definitions/Booking"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
    options:
      tags:
      - "Bookings"
      summary: "Used to respond to browser with Access-Control-Allow-Methods header.\
        \ Required for POST."
      operationId: "create_booking_options"
      parameters: []
      responses:
        "200":
          description: "Response headers will indicate the allowable methods."
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
  /bookings/{booking_id}:
    get:
      tags:
      - "Bookings"
      summary: "Retrieve Booking details by Booking ID."
      operationId: "get_booking_details"
      produces:
      - "application/json"
      parameters:
      - name: "booking_id"
        in: "path"
        description: "ID of the Booking to be retrieved."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Booking details successfully retrieved."
          schema:
            $ref: "#/definitions/Booking"
        "404":
          description: "Booking not found."
          schema:
            $ref: "#/definitions/BookingNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
    options:
      tags:
      - "Bookings"
      summary: "Used to respond to browser with Access-Control-Allow-Methods header.\
        \ Required for PATCH."
      operationId: "update_booking_options"
      parameters:
      - name: "booking_id"
        in: "path"
        description: "ID of the Booking to be updated."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Response headers will indicate the allowable methods."
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
    patch:
      tags:
      - "Bookings"
      summary: "Used to update the Status of a single Booking e.g. Cancel a Booking."
      operationId: "update_booking_status"
      consumes:
      - "application/json"
      parameters:
      - name: "booking_id"
        in: "path"
        description: "ID of the Booking to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "The patch operation to perform. Only Booking status update is\
          \ supported."
        required: true
        schema:
          $ref: "#/definitions/BookingStatusUpdate"
      responses:
        "200":
          description: "Booking status successfully updated."
          schema:
            $ref: "#/definitions/Booking"
        "404":
          description: "Booking not found."
          schema:
            $ref: "#/definitions/BookingNotFoundError"
      x-swagger-router-controller: "swagger_server.controllers.bookings_controller"
definitions:
  VenueList:
    type: "array"
    items:
      $ref: "#/definitions/Venue"
  Venue:
    type: "object"
    properties:
      venue_id:
        type: "integer"
        format: "int64"
        example: 1
      venue_name:
        type: "string"
        example: "Sydney Opera House"
      venue_desc:
        type: "string"
        example: "The Sydney Opera House is a multi-venue performing arts centre in\
          \ Sydney."
      venue_img:
        type: "string"
        example: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      venue_address:
        type: "string"
        example: "20 Smith Street, Newtown, NSW, 2000"
      seating:
        type: "array"
        items:
          $ref: "#/definitions/Seating"
    example:
      venue_name: "Sydney Opera House"
      venue_img: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      venue_address: "20 Smith Street, Newtown, NSW, 2000"
      venue_desc: "The Sydney Opera House is a multi-venue performing arts centre\
        \ in Sydney."
      venue_id: 1
      seating:
      - seating_type: "front"
        seating_number: 100
      - seating_type: "front"
        seating_number: 100
  EventList:
    type: "array"
    items:
      $ref: "#/definitions/Event"
  TicketList:
    type: "array"
    items:
      $ref: "#/definitions/Ticket"
  BookingList:
    type: "array"
    items:
      $ref: "#/definitions/Booking"
  Booking:
    type: "object"
    properties:
      booking_id:
        type: "integer"
        format: "int64"
        example: 1
      account_id:
        type: "integer"
        format: "int64"
        example: 2
      event_id:
        type: "integer"
        format: "int64"
        example: 2
      booking_status:
        type: "string"
        enum:
        - "Booked"
        - "Cancelled"
        - "Completed"
      total_cost:
        type: "number"
        format: "float"
        example: 200.0
      tickets:
        type: "array"
        items:
          $ref: "#/definitions/TicketID"
    example:
      booking_id: 1
      account_id: 2
      event_id: 2
      tickets:
      - tid: 2
      - tid: 2
      total_cost: 200.0
      booking_status: "Booked"
  Ticket:
    type: "object"
    properties:
      ticket_id:
        type: "integer"
        format: "int64"
        example: 1
      venue_id:
        type: "integer"
        format: "int64"
        example: 3
      event_id:
        type: "integer"
        format: "int64"
        example: 3
      booking_id:
        type: "integer"
        format: "int64"
        example: 5
      ticket_ref:
        type: "string"
        example: "general_1"
      ticket_status:
        type: "string"
        enum:
        - "Available"
        - "Purchased"
        - "Cancelled"
      QR_code:
        type: "string"
        example: "bc7b098c-fb66-11ec-b939-0242ac120002"
      ticket_type:
        type: "string"
        enum:
        - "General"
        - "Front"
        - "Middle"
        - "Back"
      ticket_price:
        type: "number"
        format: "float"
        example: 80.9
  Event:
    type: "object"
    properties:
      event_id:
        type: "integer"
        format: "int64"
        example: 1
      host_id:
        type: "integer"
        format: "int64"
        example: 12
      account_id:
        type: "integer"
        format: "int64"
        example: 10
      venue_id:
        type: "integer"
        format: "int64"
        example: 9
      gen_seat_price:
        type: "number"
        format: "float"
        example: 100.5
      front_seat_price:
        type: "number"
        format: "float"
        example: 80.9
      mid_seat_price:
        type: "number"
        format: "float"
        example: 70.9
      back_seat_price:
        type: "number"
        format: "float"
        example: 50.0
      event_title:
        type: "string"
        example: "Sydney KPOP Party"
      event_category:
        type: "string"
        example: "Music"
      event_short_desc:
        type: "string"
        example: "Sydney KPOP Party BTS Special!"
      event_desc:
        type: "string"
        example: "STRICTLY KPOP & K-HIPHOP! KPOP ALBUM GIVEAWAYS! LIVE DJS!"
      event_start_datetime:
        type: "string"
        example: "2015-03-25T19:00:00+10:00"
      event_end_datetime:
        type: "string"
        example: "2015-03-25T23:00:00+10:00"
      event_location:
        type: "string"
        example: "Shark Hotel Sydney, NSW"
      event_img:
        type: "string"
        example: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      event_status:
        type: "string"
        enum:
        - "Upcoming"
        - "Cancelled"
        - "Completed"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
    example:
      gen_seat_price: 100.5
      front_seat_price: 80.9
      event_end_datetime: "2015-03-25T23:00:00+10:00"
      event_location: "Shark Hotel Sydney, NSW"
      event_title: "Sydney KPOP Party"
      mid_seat_price: 70.9
      event_category: "Music"
      event_start_datetime: "2015-03-25T19:00:00+10:00"
      host_id: 12
      tags:
      - name: "Jazz"
      - name: "Jazz"
      event_short_desc: "Sydney KPOP Party BTS Special!"
      event_desc: "STRICTLY KPOP & K-HIPHOP! KPOP ALBUM GIVEAWAYS! LIVE DJS!"
      event_id: 1
      account_id: 10
      back_seat_price: 50.0
      event_img: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      venue_id: 9
      event_status: "Upcoming"
  Tag:
    type: "object"
    properties:
      name:
        type: "string"
        example: "Jazz"
    example:
      name: "Jazz"
  TicketID:
    type: "object"
    properties:
      tid:
        type: "integer"
        format: "int64"
        example: 2
    example:
      tid: 2
  Seating:
    type: "object"
    properties:
      seating_type:
        type: "string"
        example: "front"
      seating_number:
        type: "integer"
        format: "int64"
        example: 100
    example:
      seating_type: "front"
      seating_number: 100
  EventStatusUpdate:
    type: "object"
    required:
    - "op"
    - "path"
    - "value"
    properties:
      op:
        type: "string"
        example: "replace"
        enum:
        - "replace"
      path:
        type: "string"
        enum:
        - "/event_status"
      value:
        type: "string"
        example: "Cancelled"
        enum:
        - "Upcoming"
        - "Completed"
        - "Cancelled"
    example:
      op: "replace"
      path: "/event_status"
      value: "Cancelled"
  BookingStatusUpdate:
    type: "object"
    required:
    - "op"
    - "path"
    - "value"
    properties:
      op:
        type: "string"
        example: "replace"
        enum:
        - "replace"
      path:
        type: "string"
        enum:
        - "/booking_status"
      value:
        type: "string"
        example: "Cancelled"
        enum:
        - "Booked"
        - "Cancelled"
        - "Completed"
    example:
      op: "replace"
      path: "/booking_status"
      value: "Cancelled"
  Error:
    type: "object"
    properties:
      code:
        type: "string"
      type:
        type: "string"
      message:
        type: "string"
  EventNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "EventNotFoundError"
      message: "The following Event does not exist: %event_id%"
  VenueNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "VenueNotFoundError"
      message: "The following Venue does not exist: %venue_id%"
  BookingNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "BookingNotFoundError"
      message: "The following Booking does not exist: %booking_id%"
  InvalidInputError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 400
      type: "InvalidInputError"
      message: "The following mandatory fields were not provided: %field_name_1%,\
        \ %field_name_2%"
  UnexpectedServiceError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 500
      type: "UnexpectedServiceError"
      message: "The following unexpected runtime error has occurred: %error%"